+++ CHANGELOG.md
@@ -22,6 +22,7 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - Resources are updated at broker after they have been edited
 - Internal Database can be reached and viewed again at: http://localhost:8081/console
 - After editing the connector settings the broker will be updated with the new information
+- Updated recursion methods in ResourceService that caused problems
 
 ## [6.0.0] - 2021-03-26